# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "sharerr",
    "description": "The machine-facing HTTP API of a sharerr instance: the indexer feed a friend's Sonarr or Radarr queries, the BitTorrent tracker their client announces to, the gossip exchange that keeps addresses current, the lighthouse rendezvous, and the operational endpoints an orchestrator polls.\n\nThe server-rendered web UI is deliberately absent — see the `sharerr::openapi` module docs.\n\n**Authentication.** Every feed and gossip endpoint takes a peer's own API key as an `apikey` query parameter. Each friend holds a different one, so a single friend can be cut off without disturbing the others, and every request records who made it. There is no unauthenticated read path: a missing key is refused exactly the way a wrong one is, because saying \"this instance has no key configured\" would confirm the port belongs to sharerr.\n\n**Base URL.** None is listed. A sharerr instance is reached at whatever address its operator advertises — often behind a VPN with a rotating forwarded port, which is the whole reason the gossip and lighthouse endpoints exist.",
    "license": {
      "name": "AGPL-3.0-or-later"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/announce": {
      "get": {
        "tags": [
          "tracker"
        ],
        "summary": "`GET /announce`.",
        "operationId": "announce",
        "parameters": [
          {
            "name": "info_hash",
            "in": "query",
            "description": "The torrent's info hash: 20 raw bytes, percent-encoded. **Not hex.**",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "peer_id",
            "in": "query",
            "description": "The client's own id: 20 raw bytes, percent-encoded.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "port",
            "in": "query",
            "description": "The port the client accepts connections on.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            },
            "example": 6881
          },
          {
            "name": "left",
            "in": "query",
            "description": "Bytes still to download. `0` means a seeder.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "event",
            "in": "query",
            "description": "`started`, `stopped`, `completed`, or absent for a periodic re-announce.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "compact",
            "in": "query",
            "description": "`1` for the compact peer list. Anything else returns the dictionary form.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "numwant",
            "in": "query",
            "description": "How many peers to return.",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "ip",
            "in": "query",
            "description": "The client's own address. Honoured **only when it is a private address** —\notherwise the connection's source address wins, so a peer cannot announce\nsomebody else into the swarm.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A bencoded dictionary. **Always 200** — the BitTorrent tracker protocol reports refusal as a bencoded `failure reason` key, not as an HTTP status, so a client that only checks the status code will read an error as a swarm.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api": {
      "get": {
        "tags": [
          "torznab"
        ],
        "summary": "`GET /api?t=...`",
        "operationId": "torznab",
        "parameters": [
          {
            "name": "t",
            "in": "query",
            "description": "The Torznab function. `caps` for the capabilities document, or one of the\nsearch functions: `search`, `tvsearch`, `movie`, `music`, `book`.",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "tvsearch"
          },
          {
            "name": "q",
            "in": "query",
            "description": "Free-text needle, matched against the release title.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "season",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "ep",
            "in": "query",
            "description": "Daily shows send `ep=MM/DD` rather than a number. That form has no\nseason/episode to match against here, so it is read as \"no episode\nfilter\" instead of failing the whole request with a bare 400 — which\nProwlarr counts as an indexer failure and backs off from, hiding every\nother release too.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "tvdbid",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "tmdbid",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "imdbid",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A Torznab document: the capabilities XML for `t=caps`, otherwise an RSS feed of matching releases. Each item carries a `.torrent` link and a magnet whose announce tiers are attributed to the calling peer.",
            "content": {
              "application/xml": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "No such Torznab function — a Torznab `<error code=\"202\">`.",
            "content": {
              "application/xml": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "No `apikey`, or one that matches no active peer. `t=caps` requires it too, deliberately: one fewer endpoint that says anything to an unauthenticated caller.",
            "content": {
              "application/xml": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "description": "The database is not open yet.",
            "content": {
              "application/xml": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "peerApiKey": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "peerApiKey": {
        "type": "apiKey",
        "in": "query",
        "name": "apikey",
        "description": "The API key this instance issued to one friend. Issued from Settings → Peers, shown once, and stored only as a hash — losing one means issuing a new one, which is the correct behaviour for a bearer credential."
      }
    }
  },
...
}